refactor: Renamed CanvasOverlay for clarity
