+++ src/colorizer/ColorizeCanvas.ts
@@ -30,7 +30,7 @@ import vertexShader from "./shaders/colorize.vert";
 import fragmentShader from "./shaders/colorize_RGBA8U.frag";
 import pickFragmentShader from "./shaders/cellId_RGBA8U.frag";
 import Track from "./Track";
-import CanvasOverlay from "./CanvasOverlay";
+import CanvasOverlay from "./CanvasUIOverlay";
 import { FeatureThreshold } from "./types";
 
 const BACKGROUND_COLOR_DEFAULT = 0xf7f7f7;
